Chat messages + game info
Added a second info box to show the word and made the first specifically
as a timer.
The chat is now reversed and shows new messages at the bottom.
There is a function to add new messages to the chat using vanilla JS.
Made explanatory comments for variables.
Resized the size buttons to accurately depict the size to be used.
Word is auto chosen on server if none is selected in time.