Game logic + reboot
Added guessing the word and relevant information about what is happening
in the game.
The game can now be stopped and a soft reboot is possible on the server.
A few styling bits were added to format the number of characters in a
word and have the messages show appropriately.