Working size changing + Important bug fix
It is now possible to scroll over the canvas to change the size of the
paint cursor. The range in the toolbar works too and will update when
scrolling. A workaround was put in place to make sure the slider shows
alright in Edge.
A few fixes were made to make sure the right variables were used.
A REALLY important fix was made on the server to stop drawing points
from going back to the drawer, this means duplicate points aren't made.
This change will significantly speed up the system and also stops any
weird line bugs popping up such as 'spiderwebbing' or undo not working
correctly.