Choose a word!
Added an overlay to display the available words to be picked. When a
word is chosen it is sent to the server and the player gets to draw the
word.
Small fix to server to actually pick words to show on first round.
A few cleanups to style, element calls and names.