Draw clear bug fix :dash:
Fixed a nasty bug causing the clear function to set the layers and lines
to override each other and make everything explode. The layers aren't
cleared anymore, which isn't a bad thing since the color/size are saved
and there really shouldn't be any lines in their cache.
Removed the freedraw command as it isn't useful.
Moved around permissions and renamed the game.drawing property
correctly.