Nicer Buttons + Socket Fix
Buttons and the word dialog textbox actually look alright now and the
dialog now has a safety in place when it shouldn't be used.
Drawing are now dumped at the end of each turn and will be scrubbed when
the next player chooses their word.
Also fixed a bug coming from the sockets array being changed into an
object in Socket.IO 1.4.0.