Flexbox! :diamond_shape_with_a_dot_inside:
Converted to use flexbox instead of a mix of %, em and Javascript
scaling. Because of this IE9 has been dropped, although could be brought
back latter in the form of a fallback to the non-responsive version.
Also SCSS has been split into partials to make it more manageable, and
the start button has been moved into the new menu bar.