Reorganize elements + no more em sizing :eyes:
Right is always in the bottom and the left is moved in when needed. This
way text input can't be interrupted. Also added a middle breakpoint to
scale better.
Scaled back the em scaling to only the tools, everything else will scale
nomally to the pixel density (much better for mobile).
The tools are always displayed, avoiding the canvas taking the full
height when not in turn.